Make this crate no_std
As nothing actually depends on std, there is no need to keep a
dependency on it.  This allows users such as the jid crate to actually
be no_std as well.

Now that core::error::Error got stabilized, we don’t need anything from
std any longer, but this bumps the minimum supported Rust version to
1.81.0.